+++ llama_index/evaluation/faithfulness.py
@@ -96,7 +96,7 @@ class FaithfulnessEvaluator(BaseEvaluator):
         else:
             self._refine_template = refine_template or DEFAULT_REFINE_TEMPLATE
 
-    def evaluate(
+    async def aevaluate(
         self,
         query: Optional[str] = None,
         response: Optional[str] = None,
@@ -116,7 +116,7 @@ class FaithfulnessEvaluator(BaseEvaluator):
             text_qa_template=self._eval_template,
             refine_template=self._refine_template,
         )
-        response_obj = query_engine.query(response)
+        response_obj = await query_engine.aquery(response)
 
         raw_response_txt = str(response_obj)