+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Simplified portkey LLM interface (#7669)
 - Added async operation support to `ElasticsearchStore` vector store (#7613)
 - Added support for `LiteLLM` (#7600)
+- Added batch evaluation runner (#7692)
 
 ### Bug Fixes / Nits
 - Avoid `NotImplementedError` for async langchain embeddings (#7668)